Add HTML template for how the inner HTML might look like.
Write the template that includes all the information needed to
create the astronaut html from JSON.

Notes: One will need to use this template seven times to get all
astronauts. One potential issue is how template literals turn arrays
to strings. The template might need to be modified if this is not
the case.